Moved everything from EventProvider.tsx to EventContext.tsx
